Added DesktopInputVersion_Assets Compressed Folder
This zip folder contains all of the used assets and extra config files for the Desktop input version of our VR photo rendering game. Be advised that Library packages could not be included in this zip as Library folder totaled over 500 MB. Planning on making new repo or branch for Library or solely housing the Unity version projects.